feat(RCP-refactoring): refactored home page, check coin value before display
